api/requirements.txt
Django
djangorestframework
MySQL-python
